Cleanup markdown extension
* Splits symbol provider to own file
* Removes duplicate frontmatter logic
